+++ src/QuickMarcEditor/utils.js
@@ -45,7 +45,7 @@ export const isFixedFieldRow = recordRow => recordRow.tag === '008';
 export const isMaterialCharsRecord = recordRow => recordRow.tag === '006';
 export const isPhysDescriptionRecord = recordRow => recordRow.tag === '007';
 
-export const getContentSubfieldValue = (content) => {
+export const getContentSubfieldValue = (content = '') => {
   return content.split(/\$/)
     .filter(str => str.length > 0)
     .reduce((acc, str) => {
@@ -356,22 +356,14 @@ export const validateLeader = (prevLeader = '', leader = '', marcType = MARC_TYP
 };
 
 export const getLocationValue = (value) => {
-  const matches = value?.match(/\$b\s+([^$\s]+\/?)+/) || [];  
+  const matches = value?.match(/\$b\s+([^$\s]+\/?)+/) || [];
+
   return matches[0] || '';
 };
 
-export const countSubField = (field, subField) => {    
-  const reg = new RegExp('\\'+subField, "g");
-  const matches = field.content?.match(reg) || [];   
-  return matches.length;
-}
-
-export const validateSingleNoneSubfield = (field, subField) => {
-  return countSubField(field, subField)<=1;  
-}
-  
 export const validateLocationSubfield = (field, locations) => {
-  const [, locationValue] = getLocationValue(field.content)?.replace(/\s+/,' ').split(' ') || '';  
+  const [, locationValue] = getLocationValue(field.content)?.replace(/\s+/, ' ').split(' ') || '';
+
   return !!locations.find(location => location.code === locationValue);
 };
 
@@ -548,8 +540,10 @@ const validateMarcHoldingsRecord = (marcRecords, locations) => {
     return <FormattedMessage id="ui-quick-marc.record.error.location.multiple" />;
   }
 
-  if (!validateSingleNoneSubfield(marcRecords.find(({ tag }) => tag === '852'), '$b')) {    
-    return <FormattedMessage id="ui-quick-marc.record.error.field.onlyOneSubfield" values={{ fieldTag: '852', subField: '$b' }}/>;
+  if (locationRecords.length) {
+    if (getContentSubfieldValue(locationRecords[0].content).$b?.length > 1) {
+      return <FormattedMessage id="ui-quick-marc.record.error.field.onlyOneSubfield" values={{ fieldTag: '852', subField: '$b' }} />;
+    }
   }
 
   if (!validateLocationSubfield(marcRecords.find(({ tag }) => tag === '852'), locations)) {